backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
flask-socketio==5.3.2
python-socketio==5.7.2
python-engineio==4.4.0
werkzeug==2.3.7
openai-whisper==20230314     
pydub==0.25.1
transformers==4.38.0
torch>=1.12.0                
sentence-transformers==2.2.2
nltk==3.8.1
spacy==3.5.0
python-docx==0.8.11
docx2pdf==0.1.8
#sqlalchemy==2.0.19
alembic==1.11.1
python-dotenv==1.0.0
gunicorn==20.1.0
flask-socketio==5.3.2
python-socketio==5.7.2